+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - :sparkles: feat: add ControllerAdvice,ExceptionHandler
 - :bug: fix: #7 filter can't be null
 - :sparkles: feat: use ControllerAdviceExceptionResolver as default ExceptionResolver
+- refactor: refactor HandlerInterceptor
 
 
 ## v2.3.6